[수정] 에러 수정: List failed to visit cell content, returning an empty cell.
